Add TritonLauncher
